Run cron:update_safolu_from_old_amis_moedict only when 4am

+++ lib/tasks/cron.rake
@@ -66,6 +66,8 @@ namespace :cron do
   task update_safolu_from_old_amis_moedict: :environment do
     puts Time.now.to_s + "\tcron:update_safolu_from_old_amis_moedict START"
 
+    exit unless Time.now.utc.hour == 4
+
     system("rm -rf tmp/dict/s;cd /tmp;rm master.zip;rm -rf amis-moedict-master;curl -L -O https://github.com/g0v/amis-moedict/archive/refs/heads/master.zip;unzip -q master.zip 'amis-moedict-master/docs/s/*' -d .;mkdir -p /srv/web/tmp/dict/;mv amis-moedict-master/docs/s /srv/web/tmp/dict/")
 
     # 確認舊版 amis-moedict 蔡中涵大辭典的檔案數量
